refactor(): Migrate to Thor
