+++ Gemfile
@@ -3,3 +3,4 @@ source "https://rubygems.org"
 gem "roo"
 gem "elasticsearch"
 gem "rspec"
+gem "thor"
