+++ init.rb
@@ -1,0 +1,4 @@
+require "elasticsearch"
+
+# require project files
+Dir[File.dirname(__FILE__) + '/lib/*.rb'].each {|file| require file }
